Reload changes dialog, fix fulscreen on startup
You can set or unset reload changes by `:VVset [no]reloadchanged`
When it is set, if something outside changed opened files, it will show dialog and ask to keep or reload changed files. If it will keep current version, it will still prompt when you try to save it.
It sets `:set noreload` if it is on. If you unset reloadchanged, it will revert the original value of `reload`.