Reanimate from startup files errors
If you have init.vim with errors, it will not show it anywhere with --embed param and freeze nvim until user input. I send <Nul> to make it work. Errors are still not shown, can’t find the way to get it with --embed, stderr is empty, stdio does not show errors.