Show startup errors, debug mode
With --embed we can't read startup errors and it blocks process. For example, if you have error in init.vim, there is no way to get error and unblock nvim (only send <CR> to nvim until it returns to live).
With --headless we can do it, but to turn on rpc we need stdioopen that works only since nvim 0.3.
So we use --headless + stdioopen if we can. Bad luck for nvim 0.2 users if they have broken init.vim.
Also add --debug option to vv command to keep process attached and output errors to terminal.