Fix cursor position in terminal mode (#95)
Fix cursor position in terminal buffers. For some reason nvim does not send the very last cursor position in terminal buffers after redraw. If you send any command to nvim, it updates the cursor position. Added a temporary workaround, send `getMode` on `flush`. Need to investigate, it looks like a bug in nvim. I will file a ticket if it is.

Fixes #94