Setup Puppeteer tests for screen (#102)
Setup Puppeteer + jest-image-snapshot for visual regression testing for browser renderer. This is kinda heavy, requires to run dev server and headless Chrome, but there is no other reasonable way to test how it renders. This is also could be used as a base for e2e testing. But this is not the goal now.

To test renderer it starts dev server and uses Puppeteer to render the result with WebGL, canvas and everything. Then it makes a screenshot and compares it with existing snapshot. If it fails, it will save an image with diff. If this is happening in GitHub Actions, it will save these diff images as artifacts for action.

These tests are not counted in code coverage yet, but there are ways to do it (https://github.com/istanbuljs/puppeteer-to-istanbul or https://github.com/ocavue/jest-puppeteer-istanbul). Will try this separately, now I just removed `screen.ts` from code coverage report.

Also updated to new Pixi.js 6. Previous version did not work well with Puppeteer.