Fix menu bar on simple fullscreen (#105)
When you have the window with simple fullscreen, then open a new window and switch back to the previous one, status bar stays on top of the fullscreen window and does not hide as it should do.

This is a workaround to fix this bug. We have a hidden window, and when we focus on the window with simple fullscreen, we trigger setSimpleFullScreen on the hidden window and it fixes menu bar.